Move normalizeFee calculations to utilities
