+++ modules/router/src/test/services/fees.spec.ts
@@ -1,44 +1,29 @@
 import { VectorChainReader } from "@connext/vector-contracts";
-import {
-  calculateExchangeWad,
-  createTestChannelState,
-  getRandomBytes32,
-  getRandomChannelSigner,
-  getTestLoggers,
-  inverse,
-  mkAddress,
-} from "@connext/vector-utils";
-import { Result, REDUCED_GAS_PRICE, UpdateType, FullChannelState, GAS_ESTIMATES } from "@connext/vector-types";
+import * as vectorUtils from "@connext/vector-utils";
+import { Result, UpdateType, FullChannelState, GAS_ESTIMATES } from "@connext/vector-types";
 import { BigNumber } from "@ethersproject/bignumber";
-import axios from "axios";
 import { expect } from "chai";
 import Sinon from "sinon";
-import { AddressZero } from "@ethersproject/constants";
 
 import { getConfig } from "../../config";
 import * as swapService from "../../services/swap";
 import * as configService from "../../services/config";
 import { FeeError } from "../../errors";
 import * as feesService from "../../services/fees";
-import * as metrics from "../../metrics";
 
 const config = getConfig();
 
 const testName = "Router fees";
-const { log } = getTestLoggers(testName, config.logLevel ?? ("info" as any));
+const { log } = vectorUtils.getTestLoggers(testName, config.logLevel ?? ("info" as any));
 
 describe(testName, () => {
-  let coinGeckoStub: Sinon.SinonStub;
   let ethReader: Sinon.SinonStubbedInstance<VectorChainReader>;
-  let getDecimalsStub: Sinon.SinonStub;
   let getRebalanceProfileStub: Sinon.SinonStub;
   let getSwappedAmountStub: Sinon.SinonStub;
   let getFeesStub: Sinon.SinonStub;
 
   beforeEach(async () => {
-    coinGeckoStub = Sinon.stub(axios, "get");
     ethReader = Sinon.createStubInstance(VectorChainReader);
-    getDecimalsStub = Sinon.stub(metrics, "getDecimals").resolves(18);
     getRebalanceProfileStub = Sinon.stub(configService, "getRebalanceProfile");
     getSwappedAmountStub = Sinon.stub(swapService, "getSwappedAmount");
     getFeesStub = Sinon.stub(configService, "getSwapFees");
@@ -64,17 +49,17 @@ describe(testName, () => {
     beforeEach(() => {
       // default values
       transferAmount = BigNumber.from(100);
-      routerIdentifier = getRandomChannelSigner().publicIdentifier;
-      toAssetId = mkAddress("0xaaa");
-      fromAssetId = mkAddress("0xffff");
-      toChannel = createTestChannelState(UpdateType.deposit, {
-        channelAddress: mkAddress("0xaaa"),
+      routerIdentifier = vectorUtils.getRandomChannelSigner().publicIdentifier;
+      toAssetId = vectorUtils.mkAddress("0xaaa");
+      fromAssetId = vectorUtils.mkAddress("0xffff");
+      toChannel = vectorUtils.createTestChannelState(UpdateType.deposit, {
+        channelAddress: vectorUtils.mkAddress("0xaaa"),
         aliceIdentifier: routerIdentifier,
         assetIds: [toAssetId],
         networkContext: { chainId: 1 },
       }).channel;
-      fromChannel = createTestChannelState(UpdateType.deposit, {
-        channelAddress: mkAddress("0xbbb"),
+      fromChannel = vectorUtils.createTestChannelState(UpdateType.deposit, {
+        channelAddress: vectorUtils.mkAddress("0xbbb"),
         aliceIdentifier: routerIdentifier,
         assetIds: [fromAssetId],
         networkContext: { chainId: 1 },
@@ -96,8 +81,8 @@ describe(testName, () => {
 
       // by default, these functions should only return gas fee values
       // i.e. they do nothing
-      normalizedGasFeesStub = Sinon.stub(feesService, "normalizeFee");
-      normalizedGasFeesStub.onFirstCall().resolves(Result.ok(gasFees[fromChannel.channelAddress]));
+      normalizedGasFeesStub = Sinon.stub(vectorUtils, "normalizeFee");
+      normalizedGasFeesStub.resolves(Result.ok(gasFees[fromChannel.channelAddress]));
       normalizedGasFeesStub.onSecondCall().resolves(Result.ok(gasFees[toChannel.channelAddress]));
       getSwappedAmountStub.resolves(Result.ok(gasFees[toChannel.channelAddress]));
     });
@@ -302,23 +287,23 @@ describe(testName, () => {
 
     beforeEach(() => {
       toSend = BigNumber.from(100);
-      routerIdentifier = getRandomChannelSigner().publicIdentifier;
-      toAssetId = mkAddress("0xaaa");
-      fromAssetId = mkAddress("0xffff");
-      toChannel = createTestChannelState(UpdateType.deposit, {
-        channelAddress: mkAddress("0xaaa"),
+      routerIdentifier = vectorUtils.getRandomChannelSigner().publicIdentifier;
+      toAssetId = vectorUtils.mkAddress("0xaaa");
+      fromAssetId = vectorUtils.mkAddress("0xffff");
+      toChannel = vectorUtils.createTestChannelState(UpdateType.deposit, {
+        channelAddress: vectorUtils.mkAddress("0xaaa"),
         aliceIdentifier: routerIdentifier,
         assetIds: [toAssetId],
       }).channel;
-      fromChannel = createTestChannelState(UpdateType.deposit, {
-        channelAddress: mkAddress("0xbbb"),
+      fromChannel = vectorUtils.createTestChannelState(UpdateType.deposit, {
+        channelAddress: vectorUtils.mkAddress("0xbbb"),
         aliceIdentifier: routerIdentifier,
         assetIds: [fromAssetId],
       }).channel;
 
       // Stubs
-      ethReader.getCode.onFirstCall().resolves(Result.ok(getRandomBytes32()));
-      ethReader.getCode.onSecondCall().resolves(Result.ok(getRandomBytes32()));
+      ethReader.getCode.onFirstCall().resolves(Result.ok(vectorUtils.getRandomBytes32()));
+      ethReader.getCode.onSecondCall().resolves(Result.ok(vectorUtils.getRandomBytes32()));
       getRebalanceProfileStub.returns(
         Result.ok({
           chainId: 1,
@@ -332,7 +317,7 @@ describe(testName, () => {
     });
 
     it("should fail if router is not in fromChannel", async () => {
-      fromChannel.aliceIdentifier = getRandomChannelSigner().publicIdentifier;
+      fromChannel.aliceIdentifier = vectorUtils.getRandomChannelSigner().publicIdentifier;
       const result = await feesService.calculateEstimatedGasFee(
         toSend,
         toAssetId,
@@ -383,7 +368,7 @@ describe(testName, () => {
     });
 
     it("should fail if router is not in toChannel", async () => {
-      toChannel.aliceIdentifier = getRandomChannelSigner().publicIdentifier;
+      toChannel.aliceIdentifier = vectorUtils.getRandomChannelSigner().publicIdentifier;
       const result = await feesService.calculateEstimatedGasFee(
         toSend,
         toAssetId,
@@ -600,83 +585,6 @@ describe(testName, () => {
         expect(result.isError).to.be.false;
         expect(result.getValue()[toChannel.channelAddress]).to.be.eq(GAS_ESTIMATES.depositAlice);
       });
-    });
-  });
-
-  describe("feesService.normalizeFee", () => {
-    const tokenAddress = mkAddress("0xeee");
-    const chainId = 1;
-    const fee = BigNumber.from(20);
-    const exchangeRate = 15;
-
-    beforeEach(async () => {
-      coinGeckoStub.resolves({ data: { [tokenAddress]: { eth: exchangeRate } } });
-      ethReader.getGasPrice.resolves(Result.ok(REDUCED_GAS_PRICE));
-    });
-
-    it("should fail if chainId is not 1", async () => {
-      const result = await feesService.normalizeFee(fee, tokenAddress, 14, ethReader, log);
-      expect(result.isError).to.be.true;
-      expect(result.getError()?.message).to.be.eq(FeeError.reasons.ChainError);
-      expect(result.getError()?.context.message).to.be.eq("Cannot get normalize fees that are not going to mainnet");
-    });
-
-    it("should fail if no gas price override is provided && it cannot get gas price", async () => {
-      ethReader.getGasPrice.resolves(Result.fail(new Error("fail")) as any);
-      const result = await feesService.normalizeFee(fee, tokenAddress, chainId, ethReader, log);
-      expect(result.isError).to.be.true;
-      expect(result.getError()?.message).to.be.eq(FeeError.reasons.ChainError);
-      expect(result.getError()?.context.getGasPriceError).to.be.ok;
-    });
-
-    it("should fail if it cannot get decimals", async () => {
-      getDecimalsStub.rejects(new Error("Fail"));
-      const result = await feesService.normalizeFee(fee, tokenAddress, chainId, ethReader, log);
-      expect(result.isError).to.be.true;
-      expect(result.getError()?.message).to.be.eq(FeeError.reasons.ExchangeRateError);
-      expect(result.getError()?.context.message).to.be.eq("Could not get decimals");
-    });
-
-    it("should work for eth", async () => {
-      const result = await feesService.normalizeFee(fee, AddressZero, chainId, ethReader, log);
-      expect(result.isError).to.be.false;
-      expect(result.getValue()).to.be.eq(fee.mul(REDUCED_GAS_PRICE));
-    });
-
-    it("should work for tokens", async () => {
-      const result = await feesService.normalizeFee(fee, tokenAddress, chainId, ethReader, log);
-      expect(result.isError).to.be.false;
-      expect(result.getValue()).to.be.eq(
-        calculateExchangeWad(fee.mul(REDUCED_GAS_PRICE), 18, inverse(exchangeRate.toString()), 18),
-      );
-    });
-  });
-
-  describe("feesService.getExchangeRateInEth", () => {
-    const tokenAddress = config.allowedSwaps[0].fromAssetId;
-
-    it("should fail if http request fails", async () => {
-      coinGeckoStub.rejects(new Error("fail"));
-      const result = await feesService.getExchangeRateInEth(tokenAddress, log);
-      expect(result.isError).to.be.true;
-      expect(result.getError()!.message).to.be.eq(FeeError.reasons.ExchangeRateError);
-      expect(result.getError()!.context.error).to.be.eq("fail");
-    });
-
-    it("should fail if response.data[tokenAddress].eth does not exist", async () => {
-      coinGeckoStub.resolves({ data: { [tokenAddress]: {} } });
-      const result = await feesService.getExchangeRateInEth(tokenAddress, log);
-      expect(result.isError).to.be.true;
-      expect(result.getError()!.message).to.be.eq(FeeError.reasons.ExchangeRateError);
-      expect(result.getError()!.context.message).to.be.eq("Could not find rate in response");
-    });
-
-    it("should work", async () => {
-      const exchangeRate = 15;
-      coinGeckoStub.resolves({ data: { [tokenAddress]: { eth: exchangeRate } } });
-      const result = await feesService.getExchangeRateInEth(tokenAddress, log);
-      expect(result.isError).to.be.false;
-      expect(result.getValue()).to.be.eq(exchangeRate);
     });
   });
 });
