+++ modules/utils/package.json
@@ -52,6 +52,7 @@
     "@types/node": "14.14.6",
     "mocha": "8.2.0",
     "nyc": "15.1.0",
+    "sinon": "9.2.1",
     "ts-mocha": "8.0.0",
     "typescript": "4.0.5"
   }
